Fit the data using a multinomial naive Bayes classifier.
Used sklearn-pandas package to create a DataFrameMapper allowing me to add features as well as a TF-IDF vectorizer. This method looks nicer than sklearn's FeatureUnion, which isn't very elegant.

Added a feature to remove special characters and URLs from the text.

Added a feature to identify the percentage of caps locks inside a comment.